# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Taoist API
  description: |
    This is the API documentation for the Taoist platform, which provides services for both Laojun and Taishang domains.
    
    The Laojun domain focuses on plugin management, configuration management, and audit logging.
    The Taishang domain focuses on AI model management, vector collections, and task management.
  version: 1.0.0
  contact:
    name: Taoist API Support
    email: support@taoist.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.taoist.com
    description: Production server
...
paths:
...
  /api/taishang/collections:
    get:
      tags:
        - Collection
      summary: List collections
      description: Retrieves a list of all vector collections for the tenant
      operationId: listCollections
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - Collection
      summary: Create collection
      description: Creates a new vector collection
      operationId: createCollection
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CollectionCreateRequest'
      responses:
        '201':
          description: Collection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CollectionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    BaseResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "success"
        traceId:
          type: string
          example: "abc123"
      required:
        - code
        - message
        - traceId
...
    VectorCollection:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        tenant_id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174001"
        name:
          type: string
          example: "document-embeddings"
        description:
          type: string
          example: "Collection for document embeddings"
        config:
          type: object
          example: {"dimension": 1536, "metric": "cosine"}
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - tenant_id
        - name
    CollectionResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/VectorCollection'
          required:
            - data
    CollectionListResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/VectorCollection'
          required:
            - data
    CollectionCreateRequest:
      type: object
      properties:
        name:
          type: string
          example: "document-embeddings"
        description:
          type: string
          example: "Collection for document embeddings"
        config:
          type: object
          example: {"dimension": 1536, "metric": "cosine"}
      required:
        - name

    # Task schemas
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: "Invalid or expired token"
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/BaseResponse'
              - type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"